+++ flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterPlugin.kt
@@ -36,8 +36,11 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
   private var activity: WeakReference<Activity>? = null
   private var framesTracker: ActivityFramesTracker? = null
+  private var engineReadyEndtime: Long? = null
 
   override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    engineReadyEndtime = System.currentTimeMillis()
+
     context = flutterPluginBinding.applicationContext
     channel = MethodChannel(flutterPluginBinding.binaryMessenger, "sentry_flutter")
     channel.setMethodCallHandler(this)
@@ -67,6 +70,7 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
       "removeExtra" -> removeExtra(call.argument("key"), result)
       "setTag" -> setTag(call.argument("key"), call.argument("value"), result)
       "removeTag" -> removeTag(call.argument("key"), result)
+      "fetchEngineReadyEndtime" -> fetchEngineReadyEndtime(result)
       "loadContexts" -> loadContexts(result)
       else -> result.notImplemented()
     }
@@ -170,9 +174,12 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         "appStartUptimeMs" to appStartUptimeMs,
       )
 
-
       result.success(item)
     }
+  }
+
+  private fun fetchEngineReadyEndtime(result: Result) {
+    result.success(engineReadyEndtime)
   }
 
   private fun beginNativeFrames(result: Result) {