+++ flutter/lib/src/native/sentry_native_channel.dart
@@ -25,6 +25,11 @@ class SentryNativeChannel implements SentryNativeBinding {
   }
 
   @override
+  Future<int?> fetchEngineReadyEndtime() async {
+    return _channel.invokeMethod('fetchEngineReadyEndtime');
+  }
+
+  @override
   Future<void> beginNativeFrames() =>
       _channel.invokeMethod('beginNativeFrames');