+++ flutter/lib/src/sentry_flutter.dart
@@ -34,6 +34,11 @@ typedef FlutterOptionsConfiguration = FutureOr<void> Function(
 mixin SentryFlutter {
   static const _channel = MethodChannel('sentry_flutter');
 
+  /// Represents the time when the dart isolate stopped loading and is ready to execute.
+  @internal
+  // ignore: invalid_use_of_internal_member
+  static DateTime dartLoadingEnd = getUtcDateTime();
+
   static Future<void> init(
     FlutterOptionsConfiguration optionsConfiguration, {
     AppRunner? appRunner,
@@ -42,6 +47,9 @@ mixin SentryFlutter {
     @internal RendererWrapper? rendererWrapper,
   }) async {
     final flutterOptions = SentryFlutterOptions();
+
+    // ignore: invalid_use_of_internal_member
+    dartLoadingEnd = flutterOptions.clock();
 
     if (platformChecker != null) {
       flutterOptions.platformChecker = platformChecker;